refactor: move coin database components to designated packages
