+++ app/src/main/java/com/hxl/cryptonumismatist/di/DataModule.java
@@ -11,8 +11,8 @@ import com.hxl.domain.repository.CoinRepository;
 import com.hxl.domain.repository.PreferenceRepository;
 import com.hxl.local.CoinLocalImpl;
 import com.hxl.local.PreferenceLocalImpl;
-import com.hxl.local.database.BookmarkDao;
-import com.hxl.local.database.CoinDao;
+import com.hxl.local.database.coin.BookmarkDao;
+import com.hxl.local.database.coin.CoinDao;
 import com.hxl.remote.CoinRemoteImpl;
 import com.hxl.remote.api.CoinService;
 import com.hxl.remote.mapper.CoinDTOMapper;
