+++ local/src/main/java/com/hxl/local/database/coin/BookmarkDao.java
@@ -1,12 +1,12 @@
-package com.hxl.local.database;
+package com.hxl.local.database.coin;
 
 import androidx.room.Dao;
 import androidx.room.Insert;
 import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 
-import com.hxl.local.model.BookmarkEntity;
-import com.hxl.local.model.CoinEntity;
+import com.hxl.local.model.coin.BookmarkEntity;
+import com.hxl.local.model.coin.CoinEntity;
 
 import java.util.List;
